feat: [IG-11] Add automatic CHANGELOG.md
BREAKING CHANGE:

New workflow for tag releases
keywords has been change to fit with standard-version rules.

For automatic releases, commit must now start with `fix: ` or `feat: `.
No more keyword for major version, but it must be describe in the body's commit
with `BREAKING CHANGE: ` and description of the changes.